trainer/requirements.txt
torch
transformers
datasets
google-cloud-storage
